+++ apps/nextjs/src/components/pages/user/collection/index.ts
@@ -1,3 +1,2 @@
 export * from './rarity-section';
 export * from './user-sidebar';
-export * from './utils/cardUtils';
